+++ partner-matching-core/src/main/java/com/partner/partnermatch/dto/FavoriteUserDto.java
@@ -7,7 +7,7 @@ import java.util.List;
 @Data
 public class FavoriteUserDto {
     private Integer id; // 收藏记录ID
-    private Integer collectUserId; // 被收藏用户ID
+    private Long collectUserId; // 被收藏用户ID
     private String username; // 用户名
     private String avatar; // 用户头像
     private Date createdAt; // 收藏时间
