+++ partner-matching-core/src/main/java/com/partner/partnermatch/entity/UserFavorite.java
@@ -3,6 +3,7 @@ package com.partner.partnermatch.entity;
 import com.baomidou.mybatisplus.annotation.*;
 import lombok.Data;
 import java.util.Date;
+import java.util.LongSummaryStatistics;
 
 @Data
 @TableName("user_collections")
@@ -15,7 +16,7 @@ public class UserFavorite {
     private Long userId;
 
     @TableField("collect_user_id")
-    private Integer collectUserId;
+    private Long collectUserId;
 
     @TableField(value = "created_at", fill = FieldFill.INSERT)
     private Date createdAt;
